+++ src/logging_strict/tech_niques/context_locals.py
@@ -129,8 +129,8 @@ oooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooah!
 **Module private variables**
 
 .. py:data:: __all__
-   :type: tuple[str]
-   :value: ("get_locals",)
+   :type: tuple[str, str, str]
+   :value: ("get_locals", "get_locals_dynamic", "FuncWrapper")
 
    This modules exports
 
@@ -150,21 +150,13 @@ oooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooah!
 
 """
 
-from __future__ import annotations
-
+import functools
 import inspect
 import re
 import sys
 from textwrap import dedent
-from typing import (
-    Any,
-    Callable,
-    TypeVar,
-)
-from unittest.mock import (
-    MagicMock,
-    patch,
-)
+from typing import TypeVar
+from unittest.mock import patch
 
 from logging_strict.util.check_type import is_not_ok
 
@@ -173,14 +165,219 @@ if sys.version_info >= (3, 10):  # pragma: no cover
 else:  # pragma: no cover
     from typing_extensions import ParamSpec
 
-__all__ = ("get_locals",)
+__all__ = (
+    "FuncWrapper",
+    "get_locals",
+    "get_locals_dynamic",
+)
 
 _T = TypeVar("_T")  # Can be anything
 _P = ParamSpec("_P")
 
 
-def _func(param_a: str, param_b: int | None = 10) -> str:
+class FuncWrapper:
+    """Wraps a function to provide basic info about it.
+
+    :ivar func: the function to wrap
+    :vartype func: (types.FunctionType | types.MethodType | types.BuiltinFunctionType | types.BuiltinMethodType | types.WrapperDescriptorType | types.MethodWrapperType | types.MethodDescriptorType | types.ClassMethodDescriptorType)
+
+    .. seealso::
+
+       Source code source
+       https://gist.github.com/jwcompdev/65da6a59a6bcb44864de77b8a29baeed
+
+       Conversation source
+       https://stackoverflow.com/a/25959545
+
+    """
+
+    def __init__(self, func):
+        """Class constructor"""
+        if isinstance(func, functools.partial):
+            self._name = func.func.__name__
+        else:
+            self._name = func.__name__
+        self._module = inspect.getmodule(func)
+        cls = type(self)
+        self._cls = cls._get_method_parent(func)
+
+    @staticmethod
+    def _get_method_parent(meth):
+        """Returns the class of the parent of the specified method.
+
+        :param meth: the method to check
+        :returns: the class of the parent of the specified method
+        :rtype: type | None
+        """
+        # https://stackoverflow.com/users/1956611/user1956611
+        if isinstance(meth, functools.partial):
+            return FuncWrapper._get_method_parent(meth.func)
+
+        if inspect.ismethod(meth) or (
+            inspect.isbuiltin(meth)
+            and getattr(meth, "__self__", None) is not None
+            and getattr(meth.__self__, "__class__", None)
+        ):
+            for cls in inspect.getmro(meth.__self__.__class__):
+                if meth.__name__ in cls.__dict__:
+                    # bound method
+                    return cls
+            meth = getattr(meth, "__func__", meth)
+        if inspect.isfunction(meth):
+            # unbound method fallsback to __qualname__ parsing
+            cls = getattr(
+                inspect.getmodule(meth),
+                meth.__qualname__.split(".<locals>", 1)[0].rsplit(".", 1)[0],
+                None,
+            )
+            if isinstance(cls, type):
+                return cls
+        return getattr(meth, "__objclass__", None)  # handle special descriptor objects
+
+    @property
+    def name(self):
+        """Returns the function's name.
+
+        :returns: the function's name
+        :rtype: str
+        """
+        return self._name
+
+    @property
+    def cls(self):
+        """Returns the function's parent class.
+
+        :returns: the parent class
+        :rtype: type | None
+        """
+        return self._cls
+
+    @property
+    def cls_name(self):
+        """Returns the function's parent class name.
+
+        :returns: the parent class name
+        :rtype: str | None
+        """
+        if self._cls is not None:
+            ret = self._cls.__name__
+        else:
+            ret = None
+
+        return ret
+
+    @property
+    def module(self):
+        """Returns the function's parent module.
+
+        :returns: the parent module
+        :rtype: types.ModuleType | None
+        """
+        return self._module
+
+    @property
+    def module_name(self):
+        """Returns the function's parent module name.
+
+        :returns: the parent module name
+        :rtype: str | None
+        """
+        if self._module is not None:
+            ret = self._module.__name__
+        else:  # pragma: no cover
+            ret = None
+
+        return ret
+
+    @property
+    def module_filename(self):
+        """Returns the function's module filename.
+
+        :returns: the module filename
+        :rtype: str | None
+        """
+        if self._module is not None:
+            ret = self._module.__file__
+        else:
+            ret = None
+
+        return ret
+
+    @property
+    def package_name(self):
+        """Returns the function's package name.
+
+        :return: the package name
+        :rtype: str | None
+        """
+        if self._module is not None:
+            ret = self._module.__package__
+        else:
+            ret = None
+
+        return ret
+
+    @property
+    def root_package_name(self):
+        """Returns the function's root package name.
+
+        :returns: the root package name
+        :rtype: str | None
+        """
+        if self.package_name is not None:
+            ret = self.package_name.partition(".")[0]
+        else:
+            ret = None
+
+        return ret
+
+    @property
+    def full_name(self):
+        """Returns the function's full name with the class included.
+
+        :returns: the full name
+        :rtype: str
+        """
+        cls_name = self.cls_name
+        if cls_name is None:
+            ret = self.name
+        else:
+            ret = f"{self.cls_name}.{self.name}"
+
+        return ret
+
+    def get_dotted_path(self):
+        """Returns the function's full path with class and package name.
+
+        :returns: the full path
+        :rtype: str
+        :raises:
+
+        - :py:exc:`ModuleNotFoundError` -- Cannot determine func module path
+
+        """
+        cls_name = self.cls_name
+        mod_dotted_path = self.module_name
+        if mod_dotted_path is None:  # pragma: no cover
+            msg_warn = (
+                "Could not determine func module path. Undeterminable origin "
+                f"of func {self.name}"
+            )
+            raise ModuleNotFoundError(msg_warn)
+        else:
+            if cls_name is None:
+                ret = f"{mod_dotted_path}.{self.name}"
+            else:
+                ret = f"{mod_dotted_path}.{cls_name}.{self.name}"
+
+        return ret
+
+
+def _func(param_a, param_b=10):
     """Sample function to inspect the locals
+
+    Careful! In function signature, do not use ``|``. Results in
+    impossible to track down ImportError
 
     :param param_a: To whom am i speaking to? Name please
     :type param_a: str
@@ -209,7 +406,7 @@ class MockFunction:
     """Defines a Mock, for functions, to explore the execution details.
 
     :var func: Function to mock
-    :vartype func: collections.abc.Callable[..., typing.Any]
+    :vartype func: collections.abc.Callable[logging_strict.tech_niques.context_locals._P, logging_strict.tech_niques.context_locals._T]
 
     .. seealso::
 
@@ -217,16 +414,18 @@ class MockFunction:
 
     """
 
-    def __init__(self, func: Callable[..., Any]) -> None:
+    def __init__(self, func):
         """Class constructor"""
         self.func = func
+        fw = FuncWrapper(func)
+        self.full_name = fw.full_name
 
     def __call__(  # type: ignore[misc]  # missing self non-static method
-        mock_instance: MagicMock,
+        mock_instance,
         /,
-        *args: _P.args,
-        **kwargs: _P.kwargs,
-    ) -> Any:
+        *args,
+        **kwargs,
+    ):
         """Mock modifies a target function.
         The locals are included into the :paramref:`mock_instance`.
         Return value passes through
@@ -238,7 +437,7 @@ class MockFunction:
            A generic module level function. Is modified and executed. piggy backing
            the module level function ``locals`` onto the return statement
 
-        :type mock_instance: collections.abc.Callable[logging_strict.tech_niques.context_locals._P, logging_strict.tech_niques.context_locals._T]
+        :type mock_instance: unittest.mock.MagicMock
         :param args: Generic positional args
         :type args: logging_strict.tech_niques.context_locals._P.args
         :param kwargs: Generic keyword args
@@ -247,7 +446,7 @@ class MockFunction:
 
             Module levels function normal return value
 
-        :rtype: logging_strict.tech_niques.context_locals._T
+        :rtype: typing.Any
 
         .. seealso::
 
@@ -263,6 +462,7 @@ class MockFunction:
 
         # Modify to call the modified function
         # code = code + f"\nloc, ret = {mock_instance.func.__name__}(*args, **kwargs)"
+        # mock_instance.func.__name__ --> mock_instance.full_name
         support_return_tuple = (
             f"\nt_ret = {mock_instance.func.__name__}(*args, **kwargs)\n"
             "loc = t_ret[0]\n"
@@ -280,6 +480,148 @@ class MockFunction:
 
         # Return normal return value as if nothing was ever modified
         return loc["ret"]
+
+
+class MockMethod:
+    """Defines a Mock, for functions, to explore the execution details.
+
+    :var func: Function to mock
+    :vartype func: collections.abc.Callable[logging_strict.tech_niques.context_locals._P, logging_strict.tech_niques.context_locals._T]
+
+    .. seealso::
+
+       Used by :py:func:`logging_strict.tech_niques.get_locals`
+
+    """
+
+    def __init__(self, cls, func):
+        """Class constructor"""
+        self.cls = cls
+        self.func = func
+
+    def __call__(  # type: ignore[misc]  # missing self non-static method
+        mock_instance,
+        /,
+        *args,
+        **kwargs,
+    ):
+        """Mock modifies a target function.
+        The locals are included into the :paramref:`mock_instance`.
+        Return value passes through
+
+        Must use on a function with a single return value. Not yield or raise.
+
+        :param mock_instance:
+
+           A generic module level function. Is modified and executed. piggy backing
+           the module level function ``locals`` onto the return statement
+
+        :type mock_instance: unittest.mock.MagicMock
+        :param args: Generic positional args
+        :type args: logging_strict.tech_niques.context_locals._P.args
+        :param kwargs: Generic keyword args
+        :type kwargs: logging_strict.tech_niques.context_locals._P.kwargs
+        :returns:
+
+            Module levels function normal return value
+
+        :rtype: typing.Any
+
+        .. seealso::
+
+           :py:class:`typing.ParamSpec`
+
+        """
+        # Modify ``return`` statement, to include ``locals()``. Returns a tuple
+        code = re.sub(
+            "[\\s]return\\b",
+            " return locals(), ",
+            dedent(inspect.getsource(mock_instance.func)),
+        )
+
+        # Modify to call the modified function
+        # code = code + f"\nloc, ret = {mock_instance.func.__name__}(*args, **kwargs)"
+        # `usage of setattr <https://stackoverflow.com/a/18620569>`_
+        support_return_tuple = (
+            f"""setattr({mock_instance.cls.__name__}, "{mock_instance.func.__name__}", {mock_instance.func.__name__})\n"""
+            f"\nt_ret = {mock_instance.cls.__name__}.{mock_instance.func.__name__}(*args, **kwargs)\n"
+            "loc = t_ret[0]\n"
+            "ret = t_ret[1] if len(t_ret[1:]) == 1 else t_ret[1:]"
+        )
+        code += support_return_tuple
+
+        # Execute the modified function code passing in the params
+        loc = {"args": args, "kwargs": kwargs}
+        # exec(code, vars(mock_instance.module), loc)
+        exec(code, mock_instance.func.__globals__, loc)
+
+        # Put execution locals into mock instance. ``l`` is locals variable name
+        for locals_name, locals_val in loc["loc"].items():  # type: ignore[attr-defined]
+            setattr(mock_instance, locals_name, locals_val)
+
+        # Return normal return value as if nothing was ever modified
+        return loc["ret"]
+
+
+def get_locals_dynamic(
+    func,
+    /,
+    *args,
+    **kwargs,
+):
+    """Uses :py:func:`patch <unittest.mock.patch>` to retrieve the
+    tested functions locals and return value!
+
+    See this module docs for example
+
+    Limitation: the function must end with a single ``return``, not
+    ``yield`` or ``raise``.
+
+    :param func_path: dotted path to func
+    :type func_path: str
+    :param func: The func
+    :type func: collections.abc.Callable[logging_strict.tech_niques.context_locals._T, typing.Any]
+    :param args: Positional arguments
+    :type args: typing.ParamSpecArgs
+    :param kwargs: Optional (keyword) arguments
+    :type kwargs: typing.ParamSpecKwargs
+    :returns: Tuple containing return value and the locals
+    :rtype: tuple[logging_strict.tech_niques.context_locals._T, dict[str, typing.Any]]
+    """
+    fw = FuncWrapper(func)
+    # may raise ModuleNotFoundError
+    func_path_dynamic = fw.get_dotted_path()
+    if fw.cls is None:
+        # is_fcn = inspect.isfunction(func)
+        callable_func = func
+        side_effect = MockFunction(callable_func)
+    else:
+        is_has_funk = hasattr(func, "__func__")
+        if is_has_funk:
+            # staticmethod or classmethod func hidden by descriptor
+            # fw.full_name
+            # callable_func = getattr(cls, func.__name__)
+            callable_func = func.__func__
+            side_effect = MockMethod(fw.cls, callable_func)
+        else:
+            # normal method
+            callable_func = func
+            side_effect = MockMethod(fw.cls, callable_func)
+
+    with patch(
+        func_path_dynamic,
+        autospec=True,
+        side_effect=side_effect,
+    ) as mocked:
+        # mocked type is function
+        ret = mocked(*args, **kwargs)
+        # print(inspect.getmembers(mocked.side_effect))
+        d_locals = {}
+        for k, v in mocked.side_effect.__dict__.items():
+            if k != "func":
+                d_locals[k] = v
+
+        return ret, d_locals
 
 
 def get_locals(
@@ -300,13 +642,22 @@ def get_locals(
     :param func_path: dotted path to func
     :type func_path: str
     :param func: The func
-    :type func: collections.abc.Callable[..., typing.Any]
+    :type func: collections.abc.Callable[logging_strict.tech_niques.context_locals._T, typing.Any]
     :param args: Positional arguments
     :type args: typing.ParamSpecArgs
     :param kwargs: Optional (keyword) arguments
     :type kwargs: typing.ParamSpecKwargs
     :returns: Tuple containing return value and the locals
     :rtype: tuple[logging_strict.tech_niques.context_locals._T, dict[str, typing.Any]]
+
+    .. deprecated:: 1.6.0
+
+       get_locals_dynamic does not need func_path and has support for
+       class methods.
+
+       Removal not planned. Very popular, widely used, transition low
+       priority and will take time.
+
     """
     with patch(
         func_path,